fix: disable cache for injected file.
Backed merged config is inject inside one of the Javascript
static assets.
We should never allow this file to be cached on the client side,
so that any change in the configuration can be seen on the frontend
side after a backend restart.
That's particularly critical to get the latest UI configuration
of added dynamic frontend plugins.

Signed-off-by: David Festal <dfestal@redhat.com>