feat(app) an AdminPage for admin plugins (#932)
This change adds a Administration navigation tab and page intended for
plugins that are for administration of an Janus/RHDH instance.  This
commit adds two routes /admin/rbac and /admin/plugins and related
mountpoints so that dynamic plugins can contribute UI components to this
page.  The page can also deal with a few edge cases and will also not be
visible should no plugins be configured to show up on it.  A test is
added to cover these edge cases.  This commit also adds the
dynamic-plugins-info dynamic plugin.

Signed-off-by: Stan Lewis <gashcrumb@gmail.com>