feat(app): add static dynamic-plugins-info
This change adds a means to statically add and configure a frontend
plugin that needs to contribute UI elements to any of the areas of the
UI managed by the dynamic frontend implementation.  This commit also
adds the dynamic-plugins-info frontend plugin and provides a static base
configuration for it.  Finally, this change cleans up a couple of
development time warnings that get logged at the javascript console.

Signed-off-by: Stan Lewis <gashcrumb@gmail.com>