feat(app): Support extensible entity tabs (#1173)
This commit adds the ability to customize and extend the default set of
tabs available for catalog entity items.  The default set of tabs is
hard-coded in the entity page but can be reconfigured and extended per
plugin using the `entityTabs` property.  If multiple plugins target the
same entity route, only the first one will be used and a warning will be
raised.

Signed-off-by: Stan Lewis <gashcrumb@gmail.com>
Co-authored-by: David Festal <dfestal@redhat.com>