fix(ui): RHIDP-2044 enable frontend plugin loading (#1226)
* fix(ui): RHIDP-2044 enable frontend plugin loading

This change removes the authentication requirement for the scalprum
backend as it's used to serve static javascript content and related
manifests.  This commit also adds identity support to the dynamic
plugins info list frontend and migrates the backend to use the httpAuth
service.

Signed-off-by: Stan Lewis <gashcrumb@gmail.com>

* fix(ui): RHIDP-2043 add guest auth provider

This change adds the new upstream guest authentication provider to the
backend.  This provider is enabled by adding a "guest" entry to the
`auth.providers` map, like:

```yaml
auth:
  providers:
    guest: {}

auth:
  providers:
    guest:
      dangerouslyAllowOutsideDevelopment: true
```

This change also removes the dynamic-plugins-info-backend e2e test, as
it relies on an unauthenticated endpoint that isn't available anymore.
This functionaly is already covered by the UI e2e test and the
component's unit test.

Signed-off-by: Stan Lewis <gashcrumb@gmail.com>

---------

Signed-off-by: Stan Lewis <gashcrumb@gmail.com>