feat: Enable themes from dynamic plugins (#1557)
This change adds support to the frontend to load a Backstage theme
provider from a dynamic plugin.  The dynamic frontend configuration
mechanism is used to connect an exported theme provider function to an
AppTheme list which is then available when Backstage is initialized.
Dynamic plugin provided themes can either add to or replace the provided
default themes.  This change also ensures that dynamic plugins can
override the statically defined API service factories in the app as
needed.

Signed-off-by: Stan Lewis <gashcrumb@gmail.com>