fix: avoid doube rendering of menuItems backed by custom components
Issue:
When dynamicRoutes[].menuItem.importName is used to render the left-side
menuItem of a FE dynamic plugin, the item is rendered twice. Once
without the text (title), second time via the importName-component as
expected.

This fix leaves the menuItem up to be rendered via dynamicRoutesComponents
and removes it from the menuItems array.

Example config causing the issue:
dynamicPlugins:
  frontend:
    redhat.plugin-notifications:
      dynamicRoutes:
        - importName: NotificationsPage
          path: /notifications
          menuItem:
            # HERE:
            importName: NotificationsSidebarItem
            config:
              props: ...