chore(docs): fill in some missing documentation (#1852)
This change adds documentation for a few changes to the frontendi application
that have been available, namely the ability to add additional custom context
menu items to the entity page from a dynamic plugin, the ability to supply a
custom sidebar item component from a dynamic plugin and finally the
changes related to how API factories can be discovered from a frontend
dynamic plugin if it exports it's plugin object vs. the need to
explicitly configure frontend API factories.

Signed-off-by: Stan Lewis <gashcrumb@gmail.com>