Use the built-in `EnvHttpProxyAgent` from `undici`
Undici has added native support for proxy handling,
so it is no longer necessary for us to have our own custom proxy handling
for 'fetch' calls.
This was introduced in Undici v6.14.0 and enhanced in the next versions
and should fix the issues faced with the handling of the 'NO_PROXY' env var
by 'fetch'. [1][2]

Note that 'global-agent' should still continue to handle proxy settings for 'node-fetch'.

[1] https://github.com/nodejs/TSC/blob/main/meetings/2024-09-04.md?plain=1#L88-L115
[2] https://github.com/nodejs/undici/releases/tag/v6.14.0