fix(corporate-proxy): use the built-in `EnvHttpProxyAgent` from `undici` to fix the proxy issues reported (#1903)
* Use the built-in `EnvHttpProxyAgent` from `undici`

Undici has added native support for proxy handling,
so it is no longer necessary for us to have our own custom proxy handling
for 'fetch' calls.
This was introduced in Undici v6.14.0 and enhanced in the next versions
and should fix the issues faced with the handling of the 'NO_PROXY' env var
by 'fetch'. [1][2]

Note that 'global-agent' should still continue to handle proxy settings for 'node-fetch'.

[1] https://github.com/nodejs/TSC/blob/main/meetings/2024-09-04.md?plain=1#L88-L115
[2] https://github.com/nodejs/undici/releases/tag/v6.14.0

* Add some docs and reference links to understand how the settings are handled

* Set 'localhost,127.0.0.1' as default value for `NO_PROXY`

For the backend to function correctly,
requests to `localhost` need to bypass any proxy settings.
Note that defining this env var will take effect only
if the `HTTP(S)_PROXY` env vars are set.

Co-authored-by: Marcel Hild <hild@b4mad.net>

---------

Co-authored-by: Marcel Hild <hild@b4mad.net>