chore(ci): log datarouter errors in the OpenShift CI ARTIFACT_DIR
Skip PR build for changes limited to e2e folder or existing image (#2135)

Add steps to detect PRs with changes only in the `e2e/` folder and skip the build in such cases. Additionally, check if the image already exists in Quay.io and bypass the build if it does. This reduces unnecessary builds, improving efficiency in the CI workflow.